// ClientApp/src/Types/ISearchEventClassesResults.ts
export default interface ISearchEventClassesResults {
  company_advertMessage: string;
  company_companyID: string;
  company_email: string;
  company_imagesLink: string;
  company_name: string;
  company_phonenumber: string;
  company_physicalAddress: string;
  company_websiteUrlLink: string;

  eventClasses_eventClassID: string;
  eventClasses_eventID: string;
  eventClasses_classID: string;
  eventClasses_name: string;
  eventClasses_code: string;
  eventClasses_price: number;
  eventClasses_serviceFees: number;
  eventClasses_totalPrice: number;
  eventClasses_fromPlace: string;
  eventClasses_toPlace: string;
  eventClasses_fromAddress: string;
  eventClasses_toAddress: string;
  eventClasses_ticketsRangedFrom: number;
  eventClasses_ticketsRangedTo: number;
  eventClasses_totalNumberOfTickets: number;
  eventClasses_availableNumberOfTickets: number;
  eventClasses_bookedNumberOfTickets: number;
  eventClasses_isActive: boolean;
  eventClasses_activatedBy: string;
  eventClasses_activatedDate: string;
  eventClasses_deactivatedBy: string;
  eventClasses_deactivatedDate: string;
  eventClasses_descriptions: string;
  eventClasses_createdBy: string;
  eventClasses_createdDate: string;
  eventClasses_editedBy: string;
  eventClasses_editedDate: string;
  eventClasses_fromPlaceName: string;
  eventClasses_toPlaceName: string;
  event_eventID: string;
  event_productID: string;
  event_branchID: string;
  event_subcategoryID: number;
  event_name: string;
  event_code: string;
  event_fromPlace: string;
  event_toPlace: string;
  event_fromAddress: string;
  event_toAddress: string;
  event_fromDate: string;
  event_toDate: string;
  event_fromHours: string;
  event_toHours: string;
  event_descriptions: string;
}